Update README.md
# Kickstarting with Excel

## Overview of Project


### Purpose
The purpose of this analysis is to assist Louise’ in compiling and analyzing data that reveals how different campaigns fared in relation to their launch dates and funding goals. To perform this analysis, I had to create a table of data sets in excel that included the following information. Using the Kickstarter campaign data set as a reference, I applied the “COUNTIFS” function in excel to compare the number of successful, failed, and canceled plays based on funding goals. Furthermore, I was able to use the data to create a pivot table along with a chart to illustrate the information revealed from the data.  

## Analysis and Challenges

### Analysis of Outcomes Based on Launch Date

### Analysis of Outcomes Based on Goals

### Challenges and Difficulties Encountered
A challenge I faced was using the “COUNTIFS” function in excel. I had difficulties setting the goal ranges within the criteria. I was able to overcome this challenge by contacting a classmate who walked me through process and helped to identify the error in my performance of the function. 

## Results

- What are two conclusions you can draw about the Outcomes based on Launch Date?
Based on the results of the data, we can conclude that the month that launched the most successful Theater Outcome by Launch Date is May. However, June, July and August all had roughly the same number of failed launch dates. This can determine by examining the points along the trend lines of the chart.  

- What can you conclude about the Outcomes based on Goals?
Based on the result of the data we can conclude that Louie had more success in reaching her goal for plays that were less than $1000. However, she failed in reaching her goal for all plays between $45000 - $49,999.  

- What are some limitations of this dataset?

- What are some other possible tables and/or graphs that we could create?
We can use a stacked and/or clustered column chart to provide an illustration of the data set. A stacked column also gives us the ability to compare the percent Successful/Failed base on goal ranges.